+++ test/TorProcess.js
@@ -9,7 +9,7 @@ require(`${__dirname}/../src/nconf_load_env.js`)(nconf);
 nconf.defaults(require(`${__dirname}/../src/default_config.js`));
 
 describe('TorProcess', function () {
-	let tor = new TorProcess(nconf.get('torPath'), { DataDirectory: nconf.get('parentDataDirectory'), ProtocolWarnings: 0 }, null);
+	let tor = new TorProcess(nconf.get('torPath'), { Config: { DataDirectory: nconf.get('parentDataDirectory'), ProtocolWarnings: 0 } }, null);
 	describe('#create()', function () {
 		this.timeout(WAIT_FOR_CREATE);